Fix communication between clients and server
I had forgotten to create the client's IO streams, which was a pretty
big mistake..

There were also some static variables that should not have been static.
I don't think this caused any problems yet, but it would have if I
made multiple instances of the client or the server.